Changed the name and added more descrition including test for sigmoid function
